Moved button types to the types folder
