+++ components/buttons/YoutubeButton.tsx
@@ -1,8 +1,8 @@
 import { ButtonIconPosition } from '@/types/components/buttons/ButtonPropsType';
 
+import type { IButtonDefaultProps } from '../../types/components/buttons/types';
 import IconYoutube from '../icons/YouTube';
 import Button from './Button';
-import type { IButtonDefaultProps } from './types';
 
 interface IYoutubeButtonProps extends IButtonDefaultProps {}
